requirements.txt
Pillow==4.1.1
Django==1.11.2
mysqlclient==1.3.12
django-simple-captcha==0.5.5
django-sendfile==0.3.1